feature/promote-blueprint

+++ commands/utility.py
@@ -58,7 +58,12 @@ def archive(name: str, filetype: str):
 
 def promote(name: str):
     if not os.path.exists(f"ideas/forge/{name}"):
-        os.makedirs(f"ideas/forge/{name}")
-        shutil.move(f"ideas/blueprints/{name}.md", f"ideas/forge/README.md")
-        print(f"{name} promoted from blueprints to forge.")
+        if os.path.exists(f"ideas/blueprints/{name}.md"):
+            os.makedirs(f"ideas/forge/{name}")
+            shutil.move(f"ideas/blueprints/{name}.md", f"ideas/forge/README.md")
+            print(f"{name} promoted from blueprints to forge.")
+        else:
+            print(f"{name}.md doesn't exist in blueprints. Cannot promote.")
+    else:
+        print(f"{name} already exists in forge. Cannot promote.")
